AI Chat: refactor redux file
