+++ apps/src/aichat/redux/index.ts
@@ -1,0 +1,18 @@
+export * from './selectors';
+export * from './thunks';
+// Only export a subset of actions that are meant to be used by components/application code
+export {
+  AichatState,
+  clearChatMessages,
+  endSave,
+  removeUpdateMessage,
+  resetToDefaultAiCustomizations,
+  setAiCustomizationProperty,
+  setModelCardProperty,
+  setNewChatSession,
+  setShowModalType,
+  setStartingAiCustomizations,
+  setStudentChatHistory,
+  setUserHasAichatAccess,
+  setViewMode,
+} from './slice';
