+++ apps/src/aichat/redux/slice.ts
@@ -1,0 +1,286 @@
+import {createSlice, PayloadAction} from '@reduxjs/toolkit';
+
+import {registerReducers} from '@cdo/apps/redux';
+
+import {ModalTypes} from '../constants';
+import {
+  AiCustomizations,
+  ChatEvent,
+  ChatMessage,
+  FieldVisibilities,
+  LevelAichatSettings,
+  ModelCardInfo,
+  SaveType,
+  ViewMode,
+  Visibility,
+  isModelUpdate,
+  isNotification,
+  isChatMessage,
+  FeedbackValue,
+} from '../types';
+import {
+  DEFAULT_VISIBILITIES,
+  EMPTY_AI_CUSTOMIZATIONS,
+} from '../views/modelCustomization/constants';
+import {validateModelId} from '../views/modelCustomization/utils';
+
+const messageListKeys = ['chatEventsPast', 'chatEventsCurrent'] as const;
+type MessageLocation = {
+  index: number;
+  messageListKey: (typeof messageListKeys)[number];
+};
+type ModalType = ModalTypes.WARNING | ModalTypes.TEACHER_ONBOARDING | undefined;
+
+export interface AichatState {
+  // Content from previous chat sessions that we track purely for visibility to the user
+  // and do not send to the model as history.
+  chatEventsPast: ChatEvent[];
+  // Items in the current chat session that we want to provide as history to the model.
+  chatEventsCurrent: ChatEvent[];
+  // The user message currently awaiting response from the model (if any).
+  chatMessagePending?: ChatMessage;
+  // Student events viewed by a teacher user in chat workspace
+  studentChatHistory: ChatEvent[];
+  // Denotes whether we should show the warning or teacher onboarding modal
+  showModalType: ModalType;
+  initialAiCustomizations: AiCustomizations;
+  currentAiCustomizations: AiCustomizations;
+  savedAiCustomizations: AiCustomizations;
+  fieldVisibilities: FieldVisibilities;
+  viewMode: ViewMode;
+  // If a save is currently in progress
+  saveInProgress: boolean;
+  // The type of save action being performed (customization update, publish, model card save, etc).
+  currentSaveType: SaveType | undefined;
+  userHasAichatAccess: boolean;
+}
+
+const initialState: AichatState = {
+  chatEventsPast: [],
+  chatEventsCurrent: [],
+  chatMessagePending: undefined,
+  studentChatHistory: [],
+  showModalType: undefined,
+  initialAiCustomizations: EMPTY_AI_CUSTOMIZATIONS,
+  currentAiCustomizations: EMPTY_AI_CUSTOMIZATIONS,
+  savedAiCustomizations: EMPTY_AI_CUSTOMIZATIONS,
+  fieldVisibilities: DEFAULT_VISIBILITIES,
+  viewMode: ViewMode.EDIT,
+  saveInProgress: false,
+  currentSaveType: undefined,
+  userHasAichatAccess: false,
+};
+
+const aichatSlice = createSlice({
+  name: 'aichat',
+  initialState,
+  reducers: {
+    addEventToChatEventsCurrent: (state, action: PayloadAction<ChatEvent>) => {
+      state.chatEventsCurrent.push(action.payload);
+    },
+    setStudentChatHistory: (state, action: PayloadAction<ChatEvent[]>) => {
+      state.studentChatHistory = action.payload;
+    },
+    setUserHasAichatAccess: (state, action: PayloadAction<boolean>) => {
+      state.userHasAichatAccess = action.payload;
+    },
+    removeUpdateMessage: (state, action: PayloadAction<number>) => {
+      const modelUpdateMessageInfo = getUpdateMessageLocation(
+        action.payload,
+        state
+      );
+      if (!modelUpdateMessageInfo) {
+        return;
+      }
+
+      const {index, messageListKey} = modelUpdateMessageInfo;
+      state[messageListKey].splice(index, 1);
+    },
+    updateChatMessageFeedback: (
+      state,
+      action: PayloadAction<{id: number; feedback: FeedbackValue | undefined}>
+    ) => {
+      const messageToUpdate = state.studentChatHistory.find(
+        message => message.id === action.payload.id
+      );
+
+      if (messageToUpdate && isChatMessage(messageToUpdate)) {
+        messageToUpdate.teacherFeedback = action.payload.feedback;
+      }
+    },
+    clearChatMessages: state => {
+      state.chatEventsPast = [];
+      state.chatEventsCurrent = [];
+    },
+    setChatMessagePending: (state, action: PayloadAction<ChatMessage>) => {
+      state.chatMessagePending = action.payload;
+    },
+    clearChatMessagePending: state => (state.chatMessagePending = undefined),
+    setNewChatSession: state => {
+      state.chatEventsPast.push(...state.chatEventsCurrent);
+      state.chatEventsCurrent = [];
+    },
+    setShowModalType: (state, action: PayloadAction<ModalType>) => {
+      state.showModalType = action.payload;
+    },
+    setViewMode: (state, action: PayloadAction<ViewMode>) => {
+      state.viewMode = action.payload;
+    },
+    setStartingAiCustomizations: (
+      state,
+      action: PayloadAction<{
+        levelAichatSettings?: LevelAichatSettings;
+        studentAiCustomizations: AiCustomizations;
+      }>
+    ) => {
+      const {levelAichatSettings, studentAiCustomizations} = action.payload;
+
+      let reconciledAiCustomizations: AiCustomizations = {
+        ...(levelAichatSettings?.initialCustomizations ||
+          EMPTY_AI_CUSTOMIZATIONS),
+      };
+
+      for (const customizationUntyped in reconciledAiCustomizations) {
+        const customization = customizationUntyped as keyof AiCustomizations;
+
+        if (
+          (levelAichatSettings?.visibilities || DEFAULT_VISIBILITIES)[
+            customization
+          ] === Visibility.EDITABLE &&
+          studentAiCustomizations[customization]
+        ) {
+          reconciledAiCustomizations = {
+            ...reconciledAiCustomizations,
+            [customization]: studentAiCustomizations[customization],
+          };
+        }
+      }
+
+      // Make sure model ID is valid
+      reconciledAiCustomizations = {
+        ...reconciledAiCustomizations,
+        selectedModelId: validateModelId(
+          reconciledAiCustomizations.selectedModelId
+        ),
+      };
+
+      state.initialAiCustomizations = reconciledAiCustomizations;
+      state.savedAiCustomizations = reconciledAiCustomizations;
+      state.currentAiCustomizations = reconciledAiCustomizations;
+      state.fieldVisibilities =
+        levelAichatSettings?.visibilities || DEFAULT_VISIBILITIES;
+    },
+    resetToDefaultAiCustomizations: (
+      state,
+      action: PayloadAction<LevelAichatSettings | undefined>
+    ) => {
+      const levelAichatSettings = action.payload;
+
+      let defaultAiCustomizations: AiCustomizations =
+        levelAichatSettings?.initialCustomizations || EMPTY_AI_CUSTOMIZATIONS;
+
+      // Make sure model ID is valid
+      defaultAiCustomizations = {
+        ...defaultAiCustomizations,
+        selectedModelId: validateModelId(
+          defaultAiCustomizations.selectedModelId
+        ),
+      };
+
+      state.savedAiCustomizations = defaultAiCustomizations;
+      state.currentAiCustomizations = defaultAiCustomizations;
+      state.fieldVisibilities =
+        levelAichatSettings?.visibilities || DEFAULT_VISIBILITIES;
+    },
+    setSavedAiCustomizations: (
+      state,
+      action: PayloadAction<AiCustomizations>
+    ) => {
+      state.savedAiCustomizations = action.payload;
+    },
+    setAiCustomizationProperty: <T extends keyof AiCustomizations>(
+      state: AichatState,
+      action: PayloadAction<{
+        property: T;
+        value: AiCustomizations[T];
+      }>
+    ) => {
+      const {property, value} = action.payload;
+      const updatedAiCustomizations = {
+        ...state.currentAiCustomizations,
+        [property]: value,
+      };
+      state.currentAiCustomizations = updatedAiCustomizations;
+    },
+    setModelCardProperty: (
+      state,
+      action: PayloadAction<{
+        property: keyof ModelCardInfo;
+        value: ModelCardInfo[typeof property];
+      }>
+    ) => {
+      const {property, value} = action.payload;
+
+      const updatedModelCardInfo: ModelCardInfo = {
+        ...state.currentAiCustomizations.modelCardInfo,
+        [property]: value,
+      };
+      state.currentAiCustomizations.modelCardInfo = updatedModelCardInfo;
+    },
+    startSave(state, action: PayloadAction<SaveType>) {
+      state.saveInProgress = true;
+      state.currentSaveType = action.payload;
+    },
+    endSave(state) {
+      state.saveInProgress = false;
+      state.currentSaveType = undefined;
+    },
+  },
+});
+
+const getUpdateMessageLocation = (
+  id: number,
+  state: AichatState
+): MessageLocation | undefined => {
+  for (const messageListKey of messageListKeys) {
+    const messageList = state[messageListKey];
+
+    // Only allow removing individual messages that are model updates and error notifications,
+    // as we want to retain user and bot message history
+    // when requesting model responses within a chat session.
+    // If we want to clear all history
+    // and start a new session, see clearChatMessages.
+    const itemToRemovePosition = messageList.findIndex(
+      message =>
+        (isModelUpdate(message) && message.id === id) ||
+        (isNotification(message) && message.id === id)
+    );
+
+    if (itemToRemovePosition >= 0) {
+      return {index: itemToRemovePosition, messageListKey};
+    }
+  }
+};
+
+registerReducers({aichat: aichatSlice.reducer});
+
+export const {
+  addEventToChatEventsCurrent,
+  startSave,
+  setChatMessagePending,
+  clearChatMessagePending,
+  setSavedAiCustomizations,
+  updateChatMessageFeedback,
+  clearChatMessages,
+  endSave,
+  removeUpdateMessage,
+  resetToDefaultAiCustomizations,
+  setAiCustomizationProperty,
+  setModelCardProperty,
+  setNewChatSession,
+  setShowModalType,
+  setStartingAiCustomizations,
+  setStudentChatHistory,
+  setUserHasAichatAccess,
+  setViewMode,
+} = aichatSlice.actions;
