+++ apps/src/aichat/views/ChatEventsList.tsx
@@ -25,7 +25,9 @@ const ChatEventsList: React.FunctionComponent<ChatEventsListProps> = ({
 }) => {
   const [inProgrammaticScroll, setInProgrammaticScroll] = useState(false);
   const [showScrollToBottom, setShowScrollToBottom] = useState(true);
-  const {isWaitingForChatResponse} = useAppSelector(state => state.aichat);
+  const isWaitingForChatResponse = useAppSelector(
+    state => !!state.aichat.chatMessagePending
+  );
 
   const conversationContainerRef = useRef<HTMLDivElement>(null);
