+++ apps/src/aichat/views/ChatWorkspace.tsx
@@ -4,11 +4,7 @@ import Tabs, {TabsProps} from '@code-dot-org/component-library/tabs';
 import React, {useCallback, useEffect, useMemo, useState} from 'react';
 import {useSelector} from 'react-redux';
 
-import {
-  fetchStudentChatHistory,
-  selectAllVisibleMessages,
-  setShowModalType,
-} from '@cdo/apps/aichat/redux/aichatRedux';
+import {setShowModalType} from '@cdo/apps/aichat/redux';
 import TeacherOnboardingModal from '@cdo/apps/aichat/views/TeacherOnboardingModal';
 import ChatWarningModal from '@cdo/apps/aiComponentLibrary/warningModal/ChatWarningModal';
 import {useAppDispatch, useAppSelector} from '@cdo/apps/util/reduxHooks';
@@ -16,6 +12,7 @@ import {tryGetLocalStorage, trySetLocalStorage} from '@cdo/apps/utils';
 
 import {ModalTypes} from '../constants';
 import aichatI18n from '../locale';
+import {fetchStudentChatHistory, selectAllVisibleMessages} from '../redux';
 import {getShortName} from '../utils';
 
 import ChatEventsList from './ChatEventsList';
